Move model code out of example
