+++ catgrad-llm/src/lib.rs
@@ -3,6 +3,7 @@
 mod error;
 pub mod helpers;
 pub mod legacy;
+pub mod models;
 pub mod run;
 pub mod serve;
 pub mod utils;
