+++ catgrad-llm/src/models/deepseek.rs
@@ -1,9 +1,8 @@
-use crate::{Cache, llm_type};
+use crate::helpers::*;
+use crate::legacy::models::utils::Config;
 use catgrad::category::lang::eq;
 use catgrad::prelude::ops::*;
 use catgrad::prelude::*;
-use catgrad_llm::helpers::*;
-use catgrad_llm::legacy::models::utils::Config;
 use nn::*;
 
 pub struct DeepSeekModel {
