+++ catgrad-llm/src/models/gemma3.rs
@@ -1,9 +1,9 @@
-use crate::{Cache, llm_type};
+use crate::helpers::*;
+use crate::legacy::models::utils::Config;
 use catgrad::prelude::ops::*;
 use catgrad::prelude::*;
-use catgrad_llm::helpers::*;
-use catgrad_llm::legacy::models::utils::Config;
 use nn::*;
+
 pub struct Gemma3Model {
     pub config: Config,
     pub max_sequence_length: usize,
