+++ catgrad-llm/src/models/mod.rs
@@ -1,0 +1,6 @@
+pub mod deepseek;
+pub mod gemma3;
+pub mod gpt2;
+pub mod granite;
+pub mod llama;
+pub mod qwen3;
