+++ catgrad-llm/src/models/qwen3.rs
@@ -1,8 +1,7 @@
-use crate::{Cache, llm_type};
+use crate::helpers::*;
+use crate::legacy::models::utils::Config;
 use catgrad::prelude::ops::*;
 use catgrad::prelude::*;
-use catgrad_llm::helpers::*;
-use catgrad_llm::legacy::models::utils::Config;
 use nn::*;
 pub struct Qwen3Model {
     pub config: Config,
